Added tests for RLP from RFC (5,6)
