+++ testsuite/tests/kuadrant/limitador/route/test_multiple_same_rule.py
@@ -1,0 +1,26 @@
+"""Test that multiple limits targeting same rule are correctly applied"""
+
+import pytest
+
+from testsuite.gateway import RouteMatch, PathMatch, MatchType
+from testsuite.policy.rate_limit_policy import RouteSelector, Limit
+
+
+@pytest.fixture(scope="module")
+def rate_limit(rate_limit):
+    """Add limit to the policy"""
+    selector = RouteSelector(
+        RouteMatch(path=PathMatch(value="/get", type=MatchType.PATH_PREFIX)),
+    )
+    rate_limit.add_limit("test1", [Limit(8, 10)], route_selectors=[selector])
+    rate_limit.add_limit("test2", [Limit(3, 5)], route_selectors=[selector])
+    return rate_limit
+
+
+def test_two_rules_targeting_one_limit(client):
+    """Test that one limit ends up shadowing others"""
+    responses = client.get_many("/get", 3)
+    assert all(
+        r.status_code == 200 for r in responses
+    ), f"Rate Limited resource unexpectedly rejected requests {responses}"
+    assert client.get("/get").status_code == 429
