+++ testsuite/tests/kuadrant/limitador/route/test_route_rule.py
@@ -9,24 +9,9 @@ pytestmark = [pytest.mark.kuadrant_only, pytest.mark.limitador]
 
 
 @pytest.fixture(scope="module")
-def route(route, backend):
-    """Add two new rules to the route"""
-    route.remove_all_rules()
-    route.add_rule(
-        backend,
-        RouteMatch(path=PathMatch(value="/get", type=MatchType.PATH_PREFIX)),
-    )
-    route.add_rule(
-        backend,
-        RouteMatch(path=PathMatch(value="/anything", type=MatchType.PATH_PREFIX)),
-    )
-    return route
-
-
-@pytest.fixture(scope="module")
 def rate_limit(rate_limit):
     """Add limit to the policy"""
-    selector = RouteSelector(RouteMatch(path=PathMatch(value="/get", type=MatchType.PATH_PREFIX)))
+    selector = RouteSelector(RouteMatch(path=PathMatch(value="/anything", type=MatchType.PATH_PREFIX)))
     rate_limit.add_limit("multiple", [Limit(5, 10)], route_selectors=[selector])
     return rate_limit
